requirements.txt
google-generativeai
pypdf2